Back-End/requirements.txt
fastapi
uvicorn
Flask
flask-cors
Flask-Limiter
firebase-admin
requests
docker
pytz
openai
openai-agents
python-telegram-bot
discord.py
python-dotenv
waitress
PyYAML
matplotlib
firebase-admin
typing-extensions
huggingface_hub
pydantic
waitress
user-agents
stripe
